PI-48: Add workflow to generate patched tarballs
CiviCRM Core cannot be installed on sites by simply cloned the civicrm-core
repository. We actually need to download the tarball from their site and
extract the files in our project. This is due to the fact that CiviCRM is
actually a combination of files in multiple different repositories and
some additional thirdy party dependencies (like php libraries installed
with composer and frontend libs with bower and npm).

For this reason, if we ever want the patches in this fork to be applied
to a project, it's not possible to simply clone the fork. We actually
need the same files as the ones in the tarball but with our patch
applied.

This PR introduces an Actions workflow that solves this problem
by doing the following:

- Downloads the original package from the CiviCRM site
- Extracts the file
- Creates a patch file from the patches branch
- Applies this patch to the download files
- Creates a new package with the patched files
- Tag and create a new release on the fork
- Add the patched package as an asset to the release

This way, if someone ever needs to install a patched version of
CiviCRM core, all that is necessary is to download the patched tarball
from the Github Release.

The tags follow the following pattern <BASE VERSION>+patch.<PATCH VERSION>
where:

- BASE VERSION is the original version we want to patch (for example, 5.19.4)
- PATCH VERSION is the short HASH of the Git commit in the patches branch from
which the patch was generated

A limitation of this approach is that, due to the number not being sequential,
it might not be easy to see what is the latest patch for a certain version.
Making it sequential would make the workflow code more complex, so, for now,
this was not implemented. Hopefully we won't need too many patches, which
means we won't have too many tags and it will be easy to spot the latest
one simply by checking the release date on the releases page on Github.

The workflow will run automatically, whenever a new commit is pushed/merged to
the patches branch.